break out the component data json parser into a separate file
